+++ packages/components/src/components/multi-select/multi-select-option/multi-select-option.tsx
@@ -4,7 +4,6 @@ import {
   AllowedTypes,
   attachComponentCss,
   getOptionAriaAttributes,
-  getPrefixedTagNames,
   throwIfParentIsNotOfKind,
   validateProps,
 } from '../../../utils';
@@ -39,16 +38,14 @@ export class MultiSelectOption {
   public render(): JSX.Element {
     validateProps(this, propTypes);
     const { theme = 'light', selected, highlighted } = this.host;
-    attachComponentCss(this.host, getComponentCss, theme);
-
-    const PrefixedTagNames = getPrefixedTagNames(this.host);
     const isDisabled = this.disabled || this.host.disabledParent;
+
+    attachComponentCss(this.host, getComponentCss, theme, isDisabled, selected);
 
     return (
       <Host onClick={!isDisabled && this.onClick}>
         <div
           role="option"
-          aria-labelledby="checkbox"
           {...getOptionAriaAttributes(selected, isDisabled, false, !!this.value)}
           class={{
             option: true,
@@ -57,16 +54,8 @@ export class MultiSelectOption {
             'option--disabled': isDisabled,
           }}
         >
-          <PrefixedTagNames.pCheckbox
-            id="checkbox"
-            class="checkbox"
-            theme={theme}
-            checked={selected}
-            disabled={isDisabled}
-            aria-hidden="true"
-          >
-            <slot slot="label" />
-          </PrefixedTagNames.pCheckbox>
+          <span class="checkbox" aria-hidden="true" />
+          <slot slot="label" />
         </div>
       </Host>
     );